ref: create dependencies module
- move `CommonQueryParams` to dependencies
- move setup_prometheus to utils
